"Cors ItemsPage Router"

+++ PoolStoreAPI/Models/Item.cs
@@ -4,7 +4,7 @@
 namespace PoolStoreAPI.Models{
 public class Item{
  public int ItemId{get;set;}
- public decimal Cost{get;set;}
+ public float Cost{get;set;}
  public string Name{get;set;} = default!;
  public string ImageURL{get;set;} = default!;
  public int LocationId{get;set;}
